cloud_functions/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-language==1.3.0
pandas==1.0.1